requirements.txt
streamlit>=1.25.0
pdfplumber>=0.10.3
python-pptx>=0.6.21
openai>=1.14.0
python-dotenv>=1.0.1
Pillow>=10.2.0
requests>=2.31.0